Add Cecil-based component finder strategy
Mostly matches the API defined by the Reflection based finder, but uses
Mono.Cecil and its type types instead of using System.Type for everything.
With the exception of ITypeRepository (the interface is just TypeRepository
in the Reflection component finder) the type names are the same, including
namespace.

A collection of helper methods are available in the Util directory. Since
these methods mostly provide analogues to methods available on System.Type,
they are made public in order to aid anyone making use of the component
finder.