set up NuGet package metadata for projects
Example and test projects are explicitly marked to not generate packages,
while metadata for the other packages are now set up in two property groups
in the .csproj files. The first group is for setting version number and
release notes, while the second maintains general package properties for
each project.