Compare metadata token AND module MVID together for type equality checks
Metadata tokens are only guaranteed to be unique within a single module. We
need to ensure that the modules are the same when metadata tokens are the
same, and we can do that with the MVID value since it's a globally unique
value created every time a module is built.