Fix attribute detection
`CustomAttributes` collections on Cecil definition types are collections
of `CustomAttribute` and aren't castable to the actual annotation
attributes. To get around that, a set of extension methods have been
added that build and return objects of the requested attribute types, so
that we can properly locate the annotations and update the model
accordingly.

Scanning for `CodeElementAttribute` is now in its own `foreach` loop.
Using the same loop as for `ComponentAttribute` meant that the component
to which a code element was to be added might not exist when processing
the related type.

Finder functions are now properly called in `AfterFindComponents()`
instead of duplicating `FindUsedBySoftwareSystemAnnotations()` THREE
times.